fix(segfault): OpenGl context, location of kernels.cl, argument in kernel size_t/unsigned_int

+++ srcs/kernels.cl
@@ -2,7 +2,7 @@ __kernel void initSphere(
     __global float4* positions,
 	__global float4* velocities,
 	__global float4* colors,
-	const unsigned int nbParticles,
+	const unsigned long nbParticles,
 	const float radius ){
 
 }
